server: Clear request flags when we disconnect.
The server class has a bunch of request flags that remember if a request
of a certain type is in flight, this prevents making the same request
multiple times. The flags are cleared when we get a response but not
when we disconnect.

If we disconnect from the server while waiting on a response the
response will be lost and the flag won't be cleared, this will prevent
us from making any request of that type.

Most notably key queries can end up being disabled so new devices won't
be recognized anymore.