Remove list of dependencies from README
The list of dependencies in the README was out of sync with the list in requirements.txt since weechat-matrix now requires e2ee support in matrix-nio. Opening requirements.txt and installing dependencies manually is either trivial to those who know python packaging or provides no benefit to those who do not. Removing this list reduces maintenance effort and does not make installation harder to understand.