Correctly ignore markdown wrapper terminators after backslashes
The test for whether there was a following markdown wrapper terminator
(closing *, **, _, `) did not consider that there might be a backslash
before the terminator just found.
By unifying the treatment of code blocks and bold/italic wrappers, we
can handle backslashes in a uniform way too, and that is by masking them
(and the characters they escape) away when initially searching for the
final occurrence of the wrapper.