commands: Add /query support, to create a private chat with another user
Support initiating chats with other users using /query command.

When inside a matrix channel or server user id can be in the form:
 - @user:server.ext
 - user:server.ext
 - @user (it looks for @user:current-server.ext)
 - user (it looks for @user:current-server.ext)

When inside a channel, can be just the nickname of the user, and we'll
figure out the corresponding id.

Re-initiating a left chat is supported, but may need some fixes.